# storage/api-docs/art-info-api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Art-UA Info API",
        "description": "API для інформаційного порталу art-ua.info — новини, події, статті про мистецтво",
        "contact": {
            "name": "Art-UA Info Support",
            "email": "info@art-ua.info"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://art-ua.info/api/v1",
            "description": "Production Server"
        },
        {
            "url": "https://save-art.ddev.site/api/v1",
            "description": "Local Development"
        }
    ],
    "paths": {
        "/v1/catalogs": {
            "get": {
                "tags": [
                    "Catalogs"
                ],
                "summary": "Список PDF-каталогів",
                "description": "Повертає список каталогів робіт авторів з можливістю фільтрації за категорією мистецтва",
                "operationId": "getArtCatalogs",
                "parameters": [
                    {
                        "name": "art_category",
                        "in": "query",
                        "description": "Slug кореневої категорії мистецтва",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "art_subcategory",
                        "in": "query",
                        "description": "Slug підкатегорії мистецтва",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Кількість на сторінку (макс. 50)",
                        "schema": {
                            "type": "integer",
                            "default": 15,
                            "maximum": 50
                        }
                    },
                    {
                        "name": "language",
                        "in": "query",
                        "description": "Мова відповіді (uk, en)",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "uk",
                                "en"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список каталогів",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ArtCatalog"
                                            }
                                        },
                                        "links": {
                                            "$ref": "#/components/schemas/PaginationLinks"
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PaginationMeta"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/catalogs/{id}": {
            "get": {
                "tags": [
                    "Catalogs"
                ],
                "summary": "Каталог за ID",
                "description": "Повертає один PDF-каталог робіт автора",
                "operationId": "getArtCatalog",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID каталогу",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "language",
                        "in": "query",
                        "description": "Мова відповіді (uk, en)",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "uk",
                                "en"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Каталог",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/ArtCatalog"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Каталог не знайдено"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "LocalizedString": {
                "title": "LocalizedString",
                "description": "Мультимовний текст (uk/en)",
                "properties": {
                    "uk": {
                        "type": "string"
                    },
                    "en": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "PaginationMeta": {
                "title": "PaginationMeta",
                "properties": {
                    "current_page": {
                        "type": "integer"
                    },
                    "last_page": {
                        "type": "integer"
                    },
                    "per_page": {
                        "type": "integer"
                    },
                    "total": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "PaginationLinks": {
                "title": "PaginationLinks",
                "properties": {
                    "first": {
                        "type": "string",
                        "nullable": true
                    },
                    "last": {
                        "type": "string",
                        "nullable": true
                    },
                    "prev": {
                        "type": "string",
                        "nullable": true
                    },
                    "next": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "ArtCatalog": {
                "title": "ArtCatalog",
                "description": "PDF-каталог робіт автора",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "$ref": "#/components/schemas/LocalizedString"
                    },
                    "image_url": {
                        "type": "string"
                    },
                    "published_at": {
                        "type": "string",
                        "format": "date",
                        "nullable": true
                    },
                    "art_category": {
                        "properties": {
                            "slug": {
                                "type": "string"
                            },
                            "name": {
                                "$ref": "#/components/schemas/LocalizedString"
                            }
                        },
                        "type": "object",
                        "nullable": true
                    },
                    "likes_count": {
                        "type": "integer",
                        "example": 10
                    },
                    "pdf_url": {
                        "type": "string",
                        "nullable": true
                    },
                    "author": {
                        "properties": {
                            "id": {
                                "type": "integer"
                            },
                            "name": {
                                "type": "string"
                            },
                            "slug": {
                                "type": "string"
                            },
                            "avatar_url": {
                                "type": "string",
                                "nullable": true
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum Bearer Token. Формат: Bearer <token>",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            },
            "apiKey": {
                "type": "apiKey",
                "description": "API ключ для доступу до API. Передається у заголовку X-Api-Key.",
                "name": "X-Api-Key",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "apiKey": []
        }
    ]
}